% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNLR.R
\name{MNLR}
\alias{MNLR}
\title{Launch MNLR Interface}
\usage{
MNLR()
}
\value{
Nothing
}
\description{
MNLR() loads interactive user interface built using R shiny.
}
\details{
Start MNLR

The interactive user interface is to provide an easy way for people who are learning Multinomial Logistic Regression. Includes example data for testing out a few example analysis.
}
\examples{
if(interactive()){
library(rmarkdown)
MNLR()
}
}
\keyword{MNLR}
